requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
pydantic
pyyaml
pytest